pyproject.toml
[project]
name = "SteamPredictor"
version = "0.1.0"
description = "Machine learning to predict Steam metrics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "catboost>=1.2.10",
    "dotenv>=0.9.9",
    "drissionpage>=4.1.1.2",
    "fastopic>=1.0.1",
    "google-api-python-client>=2.189.0",
    "hdbscan>=0.8.42",
    "imblearn>=0.0",
    "importlib>=1.0.4",
    "ipykernel>=7.2.0",
    "langdetect>=1.0.9",
    "lxml>=6.0.2",
    "matplotlib>=3.10.8",
    "minio>=7.2.20",
    "nbformat>=5.10.4",
    "nltk>=3.9.3",
    "ollama>=0.6.1",
    "optuna>=4.8.0",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "plotly>=6.5.2",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
    "scikeras>=0.13.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.3",
    "spacy>=3.8.14",
    "statsmodels>=0.14.6",
    "stem>=1.8.2",
    "tensorflow>=2.21.0",
    "textblob>=0.19.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.3",
    "umap-learn>=0.5.11",
    "unidecode>=1.4.0",
    "wandb>=0.25.1",
    "wordcloud>=1.9.6",
    "xgboost>=3.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
